Parametrize and simplify tests Split the service logic into internal services with one responsibility per service and overall facade service publishing the external API. The facade is responsible for sanity check on inputs.
